requirements.txt
streamlit>=1.28.0
pymongo>=4.6.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
langchain>=0.1.0
openai>=1.3.0
chromadb>=0.4.15
pandas>=2.1.0
streamlit-authenticator>=0.2.3
Pillow>=10.0.0
pybase64>=1.3.0

